Move GlobalEvent to its own class.
This way, importing IGlobalEvent doesn't include it.
